Add DeveloperToolsPluginOld and assorted fixes
Adds a legacy Developer Tools plugin (DeveloperToolsPluginOld.h) and wires it into PluginBootstrap and project filters. Implements extensive tooling and UI for object/event inspection and expands GameObjects facade scanning (adds pets/special/ignored lists and Azir soldier names). Fixes AI and gameplay behavior: cache and debounce SelectTarget in AIChampionEngine, early-return when runtime spell invalid, move RefreshRuntimeSpells/PruneTrackedObjects to decision path; improve marksman ForceTarget to avoid redundant calls; adjust Camille resource checks to use float for Slider values. Orbwalker: account for lastAttackOrderToAnimGapMs in timing calculations and event handlers. Also update Azir soldier detection to use AllySpecialMinions and extend ObjectDetector scanning sources. Co-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>

+++ Nightsharp/plugins/Champion/KuroAIO/AI/Controllers/AICamilleController.h
@@ -202,7 +202,7 @@ inline bool CastQReset(const AIHeroClient& target, Mode mode, bool afterAttack) 
 inline bool CastW(const AIHeroClient& target, Mode mode, bool defensive = false) {
     if (!Engine::ValidEnemy(target, kWRange + 45.0f) || TargetCannotBeDamaged(target) ||
         !Ready(1, mode) || !Throttle(1, 90) ||
-        !HasResource(1, mode == Mode::Harass ? Slider(WMenu, "HarassReserve", 85) : 0)) return false;
+        !HasResource(1, mode == Mode::Harass ? static_cast<float>(Slider(WMenu, "HarassReserve", 85)) : 0.0f)) return false;
     const auto player = GameObjects::Player();
     if (!player.IsValid()) return false;
     const auto prediction = Engine::RuntimeSpells[1]->GetPrediction(target);
@@ -300,7 +300,7 @@ inline bool CastE(const AIHeroClient& target, Mode mode, bool fleeing = false) {
         PlannedLanding = landing;
         return true;
     }
-    if (!HasResource(2, mode == Mode::Harass ? Slider(EMenu, "HarassReserve", 110) : 0) ||
+    if (!HasResource(2, mode == Mode::Harass ? static_cast<float>(Slider(EMenu, "HarassReserve", 110)) : 0.0f) ||
         (!fleeing && TargetCannotBeDamaged(target))) return false;
     const bool lethal = !fleeing && Lethal(target,
         EDamage(target) + Q1Damage(target) + AutoDamage(target));
